+++ src/pages/Inicio/components/Preview/index.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import {  Link } from 'react-router-dom'
 
-import api from '../../../services/api'
+import api from '../../../../services/api'
 import './styles.css'
 
 function Preview(props) {
